Fix multiple installations with `solc-select use` + `--always-install` (#96)
A single version string was passed to install_artifacts instead of a version
list. This caused install_artifacts to install other versions contained
in the requested version (e.g. 0.4.2 was installed as well when 0.4.25
was requested). Call the installation function with a list as intended,
so that the correct version and that version only is installed now.

Test case:
 * `rm ~/.solc-select -rf`
 * `solc-select use 0.4.25 --always-install`

Fixes: #95